requirements.txt
discord-py~=2.5.2
python-dotenv~=1.1.0
google~=3.0.0
protobuf~=5.29.5
flask~=3.1.1
pyngrok~=7.2.11